+++ test/cbmc/proofs/MQTTPropGet_ContentType/MQTTPropGet_ContentType_harness.c
@@ -33,8 +33,8 @@ void harness()
 {
     MQTTPropBuilder_t * propBuffer;
     char ** pContentType;
-    uint16_t * contentTypeLength;
-    uint32_t currentIndex;
+    size_t * contentTypeLength;
+    size_t currentIndex;
 
     propBuffer = allocateMqttPropBuilder( NULL );
     __CPROVER_assume( isValidMqttPropBuilder( propBuffer ) );
@@ -47,7 +47,7 @@ void harness()
     }
 
     pContentType = malloc( sizeof( char * ) );
-    contentTypeLength = malloc( sizeof( uint16_t ) );
+    contentTypeLength = malloc( sizeof( size_t ) );
 
     MQTTPropGet_ContentType( propBuffer, &currentIndex, pContentType, contentTypeLength );
 }
