+++ test/cbmc/proofs/MQTT_SkipNextProperty/MQTT_SkipNextProperty_harness.c
@@ -32,7 +32,7 @@
 void harness()
 {
     MQTTPropBuilder_t * propBuffer;
-    uint32_t currentIndex;
+    size_t currentIndex;
 
     propBuffer = allocateMqttPropBuilder( NULL );
     __CPROVER_assume( isValidMqttPropBuilder( propBuffer ) );
